backend/requirements/base-requirements.txt
gunicorn==23.0.0 ; python_version >= "3.11" and python_version < "4"
django-ninja==1.3.0 ; python_version >= "3.11" and python_version < "4"
django-ninja-extra==0.22.2 ; python_version >= "3.11" and python_version < "4"
django-ninja-jwt==5.3.5 ; python_version >= "3.11" and python_version < "4"
ninja-schema==0.14.1; python_version >= "3.11" and python_version < "4"
django-environ==0.11.2 ; python_version >= "3.11" and python_version < "4"
djangorestframework==3.15.2 ; python_version >= "3.11" and python_version < "4"
djangorestframework-simplejwt==5.3.1 ; python_version >= "3.11" and python_version < "4"
djangorestframework-camel-case==1.4.2 ; python_version >= "3.11" and python_version < "4"
django-extensions==3.2.3 ; python_version >= "3.11" and python_version < "4"
drf-yasg==1.21.8 ; python_version >= "3.11" and python_version < "4"
django==5.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.14.6 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.5.3 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.4 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
django-split-settings==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
django-cachalot==2.6.3 ; python_version >= "3.11" and python_version < "4.0"
factory-boy==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==11.1.0 ; python_version >= "3.11" and python_version < "4.0"
celery[redis]==5.4.0 ; python_version >= "3.11" and python_version < "4.0"
flower==2.0.1 ; python_version >= "3.11" and python_version < "4.0"
celery_singleton==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
django-redis==5.4.0 ; python_version >= "3.11" and python_version < "4.0"
elastic-apm==6.23.0 ; python_version >= "3.11" and python_version < "4.0"
django-elasticsearch-dsl==7.4.0 ; python_version >= "3.11" and python_version < "4.0"
elasticsearch==7.10.1; python_version >= "3.11" and python_version < "4.0"
channels==4.1.0; python_version >= "3.11" and python_version < "4.0"
channels-redis==4.2.0; python_version >= "3.11" and python_version < "4.0"
djangochannelsrestframework==1.2.0; python_version >= "3.11" and python_version < "4.0"
daphne==4.1.2; python_version >= "3.11" and python_version < "4.0"